requirements.txt
pandas==2.0.3
python-dotenv==1.0.0
great-expectations==0.17.19
sqlalchemy==1.4.41
typing-extensions>=4.7.1
psycopg2-binary==2.9.7
dbt-core==1.6.1
dbt-postgres==1.6.1
pyspark==3.4.1
schedule==1.2.0
pytest==7.4.2
black==23.9.1
pylint==2.17.5
